+++ src/main/kotlin/com/github/ekvedaras/classfactoryphpstorm/domain/method/state/StateMethodReferenceOutsideFactory.kt
@@ -1,22 +1,32 @@
 package com.github.ekvedaras.classfactoryphpstorm.domain.method.state
 
 import com.github.ekvedaras.classfactoryphpstorm.domain.ClassFactory
+import com.github.ekvedaras.classfactoryphpstorm.domain.ClassFactory.Companion.asClassFactory
 import com.github.ekvedaras.classfactoryphpstorm.domain.ClassFactoryMethodReference
+import com.github.ekvedaras.classfactoryphpstorm.domain.closureState.AttributeAccess
+import com.github.ekvedaras.classfactoryphpstorm.integration.otherMethods.type.AttributesArrayValueTypeProvider
 import com.github.ekvedaras.classfactoryphpstorm.support.DomainException
 import com.github.ekvedaras.classfactoryphpstorm.support.Utilities.Companion.getActualClassReference
 import com.github.ekvedaras.classfactoryphpstorm.support.Utilities.Companion.getClass
+import com.github.ekvedaras.classfactoryphpstorm.support.Utilities.Companion.isClassFactory
 import com.github.ekvedaras.classfactoryphpstorm.support.Utilities.Companion.isClassFactoryStateMethod
 import com.intellij.psi.util.childrenOfType
+import com.jetbrains.php.lang.psi.elements.ArrayAccessExpression
 import com.jetbrains.php.lang.psi.elements.ArrayCreationExpression
 import com.jetbrains.php.lang.psi.elements.ArrayHashElement
 import com.jetbrains.php.lang.psi.elements.MethodReference
 
 class StateMethodReferenceOutsideFactory(private val methodReference: MethodReference) : ClassFactoryMethodReference {
-    override val classFactory: ClassFactory
+    override val classFactory: ClassFactory = if (methodReference.firstPsiChild is ArrayAccessExpression) {
+        val attributeAccess = AttributeAccess(methodReference.firstPsiChild as ArrayAccessExpression)
+        val type = attributeAccess.getCompleteType(AttributesArrayValueTypeProvider())
+        if (! type.isClassFactory(methodReference.project)) throw StateMethodReferenceOutsideFactoryException.notStateMethodReference()
 
-    init {
+        type.asClassFactory(methodReference.project) ?: throw StateMethodReferenceOutsideFactoryException.unableToFindMethodClass()
+    } else {
         if (!methodReference.isClassFactoryStateMethod()) throw StateMethodReferenceOutsideFactoryException.notStateMethodReference()
-        classFactory = ClassFactory(
+
+        ClassFactory(
             methodReference.getActualClassReference()?.getClass()
                 ?: throw StateMethodReferenceOutsideFactoryException.unableToFindMethodClass()
         )