+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Fixed
 - Incorrect reports of unknown columns when using nested states
 - Class property inspection when using nested states
+- Class property completion when using nested states
 
 ## [2.0.1]
 ### Fixed